LPD-54772 Remove dependencies with Structure Builder utilities in global services
