+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/components/picklist_builder/PicklistOptions.tsx
@@ -7,9 +7,8 @@ import {FrontendDataSet} from '@liferay/frontend-data-set-web';
 import {sub} from 'frontend-js-web';
 import React, {useState} from 'react';
 
+import {Option, Options} from '../../../types/Picklist';
 import {
-	Option,
-	Options,
 	useOptions,
 	useRemoveOptions,
 } from '../../contexts/PicklistBuilderContext';
