+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/types/Picklist.ts
@@ -3,6 +3,14 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
+export type Option = {
+	erc: string;
+	key: string;
+	name: Liferay.Language.LocalizedValue<string>;
+};
+
+export type Options = Map<string, Partial<Option>>;
+
 export type Picklist = {
 	externalReferenceCode: string;
 	id: number;
